move test_partial_send
